All code uploaded as of 11/07
Ignore my attempts to figure out how to use this program.

So we have individual testing programs for both the IMU and GPS along with a combined version that includes both. There's also a bluetooth testing program which means the evil stupid chip works. We had to switch to a HC-05 since the original 06 didn't work. We also have a SD testing program which so far just reads everything on the SD card.