Final Code Updates + Sanity Check
I added the # of locked on the GUI and a display that tells what port you're using so we can show transfer of data at the same time. There's also an LED sanity check code because it wasn't working in the main code before. Also added code to where the LED blinks while searching on both the teensy and led